GameLoop rudimentary version
Implemented a Wrapper for a CADisplayLink that calls the gameViewController's gameLoop. Game is basic enough (no physics) to just have a dynamic time step game loop. There is also a Game Object that takes care of updating / saving / clearing / retrieving a gameState object. gameState is saved to system defaults using a JSON Encoder/Decoder. ArrowView is using a custom padding view to display a slightly bigger background than the arrow size. "Running" arrows need to be syncronized with the "goalArrows". the gameViewController uses app Notifications in order to pause the game while the app is inactive. Beats are syncronized with the game speed and the framerate.