Added Swipe Gestures. Gameplay finished
Game now supports swipe gesture and is playable though it does not have an ending (yet). Added Goal Letters using AutoLayout and custom OutlinedLabel. Have fixed the arrow size when game starts and it remains the same for all Orientations. Updating game state now receives the number of arrows that fit (vertically) on the game screen.