Custom DnaView
Created a Custom DnaView that draws itself in real time. The double helix is composed of pair segments (with circles at extremities) and there is a torsion applied to each segment pair. Several gestures control the view: pinch - scale or torsion, scroll (parent scrollVIew), pan - 3D rotation and finally touch (which is not implemented yet).  Initial View Controller is set to test this new dnaView.