DnaView precise mapSync
Changes in dnaView bounds would call updateDimensions() which in turn would call setNeedsLayout() and change the bounds thus creating an infinite loop. Initially the problem was solved by using a Bool indicator. This updates removes that Boolean indicator and relies on the 2 parts calling each other until the internal height of the helix is not changing. This provides far better results and perfect synconization with the 'syncMaView'